% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariances.R
\name{covariance_p}
\alias{covariance_p}
\title{Computes covariance of p}
\usage{
covariance_p(p, n)
}
\arguments{
\item{p}{proportion of animals that died at each observed time.}

\item{n}{total number of animals.}
}
\description{
Computes covariance of p
}
