% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariances.R
\name{covariance_p_corrected}
\alias{covariance_p_corrected}
\title{covariance corrected}
\usage{
covariance_p_corrected(pcont, pexp, ncont, nexp)
}
\arguments{
\item{pcont, pexp}{proportion of animals died at each observed time in the control
and experimental group respectively.}

\item{ncont, nexp}{total number of animales in the control and
experimental group respectively.}
}
\description{
covariance corrected
}
