% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_linear}
\alias{fit_linear}
\title{Linear model with matrix weight}
\usage{
fit_linear(y, x, cov)
}
\arguments{
\item{y}{independent variable}

\item{x}{dependent variable}

\item{cov}{covariance matrix}
}
\description{
Linear model with matrix weight
}
