% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letha-dose.R
\name{ld_fit}
\alias{ld_fit}
\title{Fits a logit}
\usage{
ld_fit(dose, dead, n = max(dead), trans_y = "logit")
}
\arguments{
\item{dose}{Vector of doses for each experiment.}

\item{dead}{Vector of acummulated number of animals dead on each experiment.}

\item{n}{Total number of animals in each experimnet. By default is the maximum number
of dead animals of all experiments.}

\item{trans_y}{A string describing how to transform the data. Options are
\code{"logit"}, \code{"probit"} and \code{"cll"} (for the Complementary log-log transformation).
(Note that \code{"identiy"} is not supported).}
}
\description{
Fits a logit
}
\examples{

data <- data.frame(dose = c(38,40,40,40,42,42,42,44,44,44,47,47,47),
                   dead = c(0,0,0,0,1,5,5,15,17,18,20,20,20))

model <- with(data, ld_fit(dose, dead, n = 20))
lethal_dose(model, p = c(0.25, 0.5, .90))

plot(model, p = c(0.25, 0.5, 0.9))

}
