% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letha-dose.R
\name{lethal_dose}
\alias{lethal_dose}
\title{Compute lethal dose}
\usage{
lethal_dose(model, p = 0.5, alpha = 0.05)
}
\arguments{
\item{model}{A model returned by \link{ld_fit}}

\item{p}{Proportion of dead animals.}

\item{alpha}{Confidence level.}
}
\description{
Compute lethal dose
}
