% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence-interval.R
\name{lethal_time}
\alias{lethal_time}
\title{Computes Lethal Time and it's confidence interval}
\usage{
lethal_time(model, p = 0.5, alpha = 0.05)
}
\arguments{
\item{model}{the linear model fitter with \link{fit_linear}}

\item{p}{the proportion of animals dead.}

\item{alpha}{confidence level}
}
\description{
Uses Fieller's theorem.
}
