% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{lt_fit}
\alias{lt_fit}
\title{Fit probit}
\usage{
lt_fit(
  times,
  dead_experiment,
  dead_control,
  n_experiment = max(dead_experiment),
  n_control = n_experiment,
  trans_y = "logit",
  tol = 0.001,
  max_iter = 50
)
}
\arguments{
\item{times}{A vector of times in which observations took place.}

\item{dead_experiment, dead_control}{Acummulated number of dead animals in the
treatment group and control group, respectively.}

\item{n_experiment}{Total number of animals in the treatment group. The
default is the maximum number of \code{dead_experiment}.}

\item{n_control}{Total number of animals in the control group. The default
is the same number of animals in treatment group.}

\item{trans_y}{A string describing how to transform the data. Options are
\code{"logit"}, \code{"probit"},  \code{"cll"} (for the Complementary log-log transformation) and
\code{"identity"} (i.e. do not transform the data).}

\item{tol, max_iter}{Relative tolerance and maximum number of iterations
for the iterative fit.}
}
\description{
Fits a probit (or other transformations) to mortality data
}
\examples{

# Some hipothetical data
experiment <- data.frame(
  days = c(0L,1L,2L,3L,4L,5L,6L,7L,
           8L,9L,10L,11L,12L,13L,14L),
  control = c(0L,0L,0L,3L,3L,5L,5L,5L,
              5L,5L,6L,6L,6L,6L,6L),
  treatment_1 = c(0L,13L,23L,28L,29L,29L,30L,
                  30L,30L,30L,30L,30L,30L,30L,30L),
  treatment_2 = c(0L,10L,24L,28L,29L,29L,29L,
                  29L,30L,30L,30L,30L,30L,30L,30L))

 model <- with(experiment, lt_fit(days, treatment_1, control))

 lethal_time(model)

 plot(model)


}
\references{
Pro bit Analysis of Correlated Data: Multiple Observations Over TiIne at One Concentration
Code adapted from https://www.ars.usda.gov/pacific-west-area/parlier/sjvasc/cpq/docs/probit-download/
}
