% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{predict.killtime_lt_fit}
\alias{predict.killtime_lt_fit}
\title{Predit new values from a fit}
\usage{
\method{predict}{killtime_lt_fit}(object, newdata = NULL, interval = FALSE, alpha = 0.05)
}
\arguments{
\item{object}{A model returned by \link{lt_fit}.}

\item{newdata}{A vector of times. If omitted, the fitted values are used.}

\item{interval}{Whether to compute confidence interval.}

\item{alpha}{Confidence level for interval calculation.}
}
\description{
Predit new values from a fit
}
